requirements.txt
spotipy>=2.23.0
SQLAlchemy>=1.4.39
requests>=2.25.1
psycopg2>=2.9.1
pandas>=2.0.1
numpy>=1.24.3
python-dotenv>=1.0.0